separate PositioningData class into its own file
